Split function to new file to prevent dev node loading issue with circular references
