+++ src/Steps/Distribution/Fields/DSO.tsx
@@ -3,7 +3,7 @@ import { CAFTOPDistribution } from "api/CAFTOP";
 import { Radio } from "@fluentui/react-components";
 import BACRadioGroup from "components/BaseFormFields/BACRadioGroup";
 import { Distribution } from "stateManagement/reducer";
-import { populateWithDefaultValue } from "utilities/Validations";
+import { populateWithDefaultValue } from "utilities/Validation";
 
 /** Rule for when the DSO is Not Applicable */
 export const dsoRuleNA = z.object({
