+++ src/utilities/Validations.ts
@@ -57,14 +57,6 @@ import {
   outsidedsoRuleSave,
 } from "Steps/Distribution/Fields/OutsideDSO";
 
-/** Funciton to replace the value in the data with the provided value
- * @param value The value which will replace the value in the data
- * @returns A ZodEffects object that transform the value from what it was to the value provided
- */
-export const populateWithDefaultValue = <T>(value: T) => {
-  return z.any().transform((_obj) => value);
-};
-
 const useAddlPECValidation = (schema: ZodSchema<CAFTOPInfo>) => {
   const ProgramNamesAndECs = useProgramNamesAndECs();
   return schema.superRefine((data, ctx) => {
